+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+### Added
+- EmailAddress class.
 
 ## 1.0.0 - 2017-05-23
 - First stable revision.